Move simplePatternToRegex to psi-utils
