+++ detekt-rules-libraries/build.gradle.kts
@@ -4,9 +4,11 @@ plugins {
 
 dependencies {
     compileOnly(projects.detektApi)
+    compileOnly(projects.detektPsiUtils)
 
     testImplementation(libs.kotlin.compiler)
     testImplementation(projects.detektApi)
+    testRuntimeOnly(projects.detektPsiUtils)
     testImplementation(projects.detektTest)
     testImplementation(projects.detektTestAssertj)
     testImplementation(projects.detektTestJunit)
