+++ detekt-rules-naming/src/main/kotlin/dev/detekt/rules/naming/ForbiddenClassName.kt
@@ -6,7 +6,7 @@ import dev.detekt.api.Entity
 import dev.detekt.api.Finding
 import dev.detekt.api.Rule
 import dev.detekt.api.config
-import dev.detekt.api.simplePatternToRegex
+import dev.detekt.psi.simplePatternToRegex
 import org.jetbrains.kotlin.psi.KtClassOrObject
 
 /**
